# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/workout/routines": {
      "post": {
        "description": "Create a new routine. Any authenticated user can create routines. Routines created by ADMIN users are automatically marked as official.",
        "operationId": "RoutinesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Routine created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Routine created successfully",
                    "data": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Upper/Lower Hypertrophy",
                      "description": "4-day program focused on strength and hypertrophy.",
                      "exerciseIds": [
                        "123e4567-e89b-12d3-a456-426614174001"
                      ],
                      "routineGoal": [
                        "muscle_gain"
                      ],
                      "official": false,
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-01T12:00:00Z",
                      "nExercises": 1,
                      "userId": "123e4567-e89b-12d3-a456-426614174099"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error. Invalid or missing required fields",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "name must be longer than or equal to 3 characters",
                      "description must be longer than or equal to 10 characters",
                      "exerciseIds must contain at least 1 elements",
                      "routineGoal must contain at least 1 elements"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create routine",
        "tags": [
          "Routines"
        ]
      },
      "get": {
        "description": "Get a paginated list of all routines. Accessible by all authenticated users. This action updates the user daily streak.",
        "operationId": "RoutinesController_findAll",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of items to skip",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated routine list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Routines retrieved successfully",
                    "data": {
                      "items": [
                        {
                          "id": "123e4567-e89b-12d3-a456-426614174000",
                          "name": "Upper/Lower Hypertrophy",
                          "description": "4-day program focused on strength and hypertrophy.",
                          "routineGoal": [
                            "muscle_gain"
                          ],
                          "official": false,
                          "exercises": [
                            {
                              "id": "123e4567-e89b-12d3-a456-426614174001",
                              "name": "Barbell Bench Press",
                              "description": "Composite exercise for chest development."
                            }
                          ],
                          "nExercises": 1,
                          "createdAt": "2024-01-01T12:00:00Z",
                          "updatedAt": "2024-01-01T12:00:00Z"
                        }
                      ],
                      "total": 30,
                      "limit": 10,
                      "offset": 0
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List routines with pagination",
        "tags": [
          "Routines"
        ]
      }
    },
    "/workout/routines/{id}": {
      "get": {
        "description": "Retrieve a specific routine by its ID. Accessible by all authenticated users.",
        "operationId": "RoutinesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Routine UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Routine retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Routine retrieved successfully",
                    "data": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Upper/Lower Hypertrophy",
                      "description": "Description...",
                      "routineGoal": [
                        "muscle_gain"
                      ],
                      "official": false,
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-01T12:00:00Z",
                      "nExercises": 5,
                      "exercises": [
                        {
                          "id": "123e4567-e89b-12d3-a456-426614174001",
                          "name": "Bench Press"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          },
          "404": {
            "description": "Routine not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get routine by ID",
        "tags": [
          "Routines"
        ]
      },
      "patch": {
        "description": "Update a routine. Only the owner of the routine or users with ADMIN role can update it.",
        "operationId": "RoutinesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Routine UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Routine updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "message": "Routine updated successfully",
                    "data": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Upper/Lower Hypertrophy",
                      "description": "Updated description...",
                      "routineGoal": [
                        "muscle_gain"
                      ],
                      "official": false,
                      "createdAt": "2024-01-01T12:00:00Z",
                      "updatedAt": "2024-01-02T12:00:00Z",
                      "nExercises": 5
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Validation error. Invalid field values",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "name must be longer than or equal to 3 characters",
                      "exerciseIds must contain at least 1 elements"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          },
          "403": {
            "description": "Forbidden - You can only modify routines that belong to you"
          },
          "404": {
            "description": "Routine not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update routine",
        "tags": [
          "Routines"
        ]
      },
      "delete": {
        "description": "Delete a routine. Only the owner of the routine or users with ADMIN role can delete it.",
        "operationId": "RoutinesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Routine UUID",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Routine deleted successfully"
          },
          "401": {
            "description": "Unauthorized - No valid authentication token"
          },
          "403": {
            "description": "Forbidden - You can only delete routines that belong to you"
          },
          "404": {
            "description": "Routine not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete routine",
        "tags": [
          "Routines"
        ]
      }
    },
...
  },
  "info": {
    "title": "AthletIA API",
    "description": "API documentation for AthletIA",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RoutineRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Routine name",
            "minLength": 3,
            "maxLength": 50,
            "example": "Upper/Lower Hypertrophy"
          },
          "description": {
            "type": "string",
            "description": "Routine description",
            "minLength": 10,
            "maxLength": 500,
            "example": "4-day program focused on strength and hypertrophy."
          },
          "exerciseIds": {
            "description": "Associated exercise IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "routineGoal": {
            "type": "array",
            "description": "Routine goals - Valid values: weight_loss, muscle_gain, weight_maintenance, endurance, flexibility, general_fitness, rehabilitation, improved_posture, balance_and_coordination, cardiovascular_health, strength_training, athletic_performance, lifestyle_enhancement",
            "items": {
              "type": "string",
              "enum": [
                "weight_loss",
                "muscle_gain",
                "weight_maintenance",
                "endurance",
                "flexibility",
                "general_fitness",
                "rehabilitation",
                "improved_posture",
                "balance_and_coordination",
                "cardiovascular_health",
                "strength_training",
                "athletic_performance",
                "lifestyle_enhancement"
              ]
            }
          },
          "official": {
            "type": "boolean",
            "description": "Official routine flag",
            "example": false
          }
        },
        "required": [
          "name",
          "description",
          "routineGoal",
          "official"
        ]
      },
      "RoutineUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Routine name",
            "minLength": 3,
            "maxLength": 50,
            "example": "Upper/Lower Hypertrophy"
          },
          "description": {
            "type": "string",
            "description": "Routine description",
            "minLength": 10,
            "maxLength": 500
          },
          "exerciseIds": {
            "description": "Associated exercise IDs",
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "routineGoal": {
            "type": "array",
            "description": "Routine goals",
            "items": {
              "type": "string",
              "enum": [
                "weight_loss",
                "muscle_gain",
                "weight_maintenance",
                "endurance",
                "flexibility",
                "general_fitness",
                "rehabilitation",
                "improved_posture",
                "balance_and_coordination",
                "cardiovascular_health",
                "strength_training",
                "athletic_performance",
                "lifestyle_enhancement"
              ]
            }
          },
          "official": {
            "type": "boolean",
            "description": "Official routine flag",
            "example": false
          }
        }
      },
...
    }
  }
}